Move Author model to Blogging\Models namespace
