+++ config/hyde.php
@@ -11,9 +11,9 @@
 |
 */
 
+use Hyde\Blogging\Models\Author;
 use Hyde\Framework\Helpers\Features;
 use Hyde\Framework\Helpers\Meta;
-use Hyde\Framework\Models\Support\Author;
 
 return [
     /*
