+++ packages/framework/src/Blogging/Models/Author.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Models\Support;
+namespace Hyde\Blogging\Models;
 
 use Illuminate\Support\Collection;
 use Stringable;
