+++ packages/framework/src/Framework/Concerns/Internal/ConstructsPageSchemas.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Concerns\Internal;
 
+use Hyde\Blogging\Models\Author;
 use Hyde\Framework\Actions\Constructors\FindsNavigationDataForPage;
 use Hyde\Framework\Actions\Constructors\FindsTitleForPage;
 use Hyde\Framework\Contracts\FrontMatter\BlogPostSchema;
 use Hyde\Framework\Models\Pages\MarkdownPost;
-use Hyde\Framework\Models\Support\Author;
 use Hyde\Framework\Models\Support\DateString;
 use Hyde\Framework\Models\Support\Image;
 use Hyde\Hyde;
